### 2018-01-23 #187 @DomBennett
phylotaR -- submission
*R implementation of the [PhyLoTa](http://phylota.net/) pipeline. It identifies orthologous sequences in GenBank using BLAST searches. It consists of three automated stages: taxise, download and cluster.*

```
Package: phylotaR
Title: Automated phylogenetic sequence cluster identification from GenBank
Version: 0.1
Author: Hannes Hettling, Rutger Vos, Alexander Zika, Dominic J. Bennett, Alexandre Antonelli
Maintainer: D.J. Bennett <dominic.john.bennett@gmail.com>
Description: PhyLoTa identifies orthologous sequence clusters from data readily available in GenBank. This is an R implementation of this process.
SystemRequirements: BLAST+ (>2.7.1)
    R (>= 3.3.0),
    foreach
    curl,
    doMC,
    doSNOW,
    igraph,
    CHNOSZ,
    rentrez,
    DBI,
    taxize,
    XML
```

-   URL for the package (the development repository, not a stylized html page): https://github.com/DomBennett/phylotaR

*data retrieval, it downloads sequences from genbank and identifies sequence clusters that may be of use for phylogenetic analysis*

*Evolutionary biologists, phylogeneticists. Determining sequence orthology is the essential first step of any phylogenetic analysis. Usually a phylogeneticist may search GenBank for sequences of the same gene, download and align. Due to naming problems, however, potential relevant sequences may be missed or sequences that have been mislabelled are downloaded. PhyLoTa gets around this by performing all-v-all BLAST searches and does not rely on gene names. PhyLoTa is often the first step of any large-scale automated phylogenetic pipeline (e.g. [SUPERSMART](http://www.supersmart-project.org/)). Potentially, this R package could be central to many future phylogenetic analyses.*

*The only current approach for identifying PhyLoTa clusters is the PhyLoTa browser. The data underlying its current release, however, is five years old. In those five years there have been [over 40 million new sequences](https://www.ncbi.nlm.nih.gov/genbank/statistics/) added to GenBank. An alternative is needed to make use of the latest data. Additionally, phylotaR would allow a user to specify their own parameters when running the pipeline.*

	- [ ] You intend to archive the code for the package in a long-term repository which meets the requirements of the journal.

*No errors, but plenty of warnings to do with documentation.*

*Not just yet. Outstanding: conversion to snake_case, generate README from rmd, create website with pkgdown, NEWS required, authors@R syntax, depends to imports and cat to message (although is that necessary given the logging methods?)*

## Pre-submission Note

*Hi!
I've produced this pre-submission issue to determine whether phylotaR would be something you'd like added to your roster? On our side, we would be very keen as we think future phylogenetic analyses will depend on making software open-source, modular and collaborative. I did feel, however, that it may not quite fit into the ropensci family, because it's a pipeline and it depends on software external to R (BLAST). Please be aware that phylotaR is still only being developed and a lot of testing still needs to be performed. If you are interested, I'd be happy to update the naming style (I prefer camelcase), improve the documentation and whatever else is needed to meet your criteria.
Dom*
