### 2018-01-30 #187 @sckott
### Editor checks:

---

#### Editor comments

Thanks for your submission @DomBennett   !!

- You didn't check the top level MEE box - i've checked that box. You didn't check the box for `The manuscript describing the package is no longer than 3000 words` - will manuscript length be a problem?
- You didn't check the box for `has a CRAN and OSI accepted license.` I think you do, can you check that box?
- You didn't check the box for `does not violate the Terms of Service of any service it interacts with.` Is that going to be a problem? Can you check on that?
- I see you checked the submit to MEE box. Note that we can not make any gaurantee about MEE considering your paper a good fit for them, or them accepting the paper
- Seems like `blast` on Homebrew (installed via `brew install blast`) is the same as at the NCBI link you gave at https://www.ncbi.nlm.nih.gov/books/NBK279671/  - Is that right? homebrew is for OSX platforms only
- It would indeed be ideal if the pkg didn't have to use system calls to external tools, but there are times when there's no other way.
- In terms of style, we prefer snake case but as long as you are consistent within the package that is THE most important thing (i.e, don't use a combination of camelCase, TitleCase and snake_case).
- You did not check the box for `Does the package conform to rOpenSci packaging guidelines?` Can you do that and describe any exceptions
- You can put a ropensci review badge in your README

- Goodpractice output for you and reviewers to consider:

```r
── GP phylotaR ────────────

It is good practice to

    R/base-wrappers.R:7:NA
    R/blast-tools.R:10:NA
    R/blast-tools.R:16:NA
    R/blast-tools.R:28:NA
    R/blast-tools.R:34:NA
    ... and 325 more lines

  ✖ not use "Depends" in DESCRIPTION, as it can cause name
    clashes, and poor interaction with other packages. Use "Imports"
    instead.
  ✖ omit "Date" in DESCRIPTION. It is not required and it
    gets invalid quite often. A build date will be added to the package
    when you perform `R CMD build` on it.
  ✖ use '<-' for assignment instead of '='. '<-' is the
    standard, and R users and developers are used it and it is easier
    to read your code for them if you use '<-'.

    R/ncbi-tools.R:156:9

    R/blast-tools.R:59:1
    R/taxise-tools.R:361:1

  ✖ omit trailing semicolons from code lines. They are not
    needed and most R coding standards forbid them

    R/ncbi-tools.R:147:55

    vapply() instead.

    R/blast-tools.R:32:13
    R/blast-tools.R:33:11
    R/cache-tools.R:253:13
    R/cache-tools.R:256:13
    R/cache-tools.R:317:12
    ... and 32 more lines

  ✖ avoid 1:length(...), 1:nrow(...), 1:ncol(...),
    1:NROW(...) and 1:NCOL(...) expressions. They are error prone and
    result 1:0 if the expression on the right hand side is zero. Use
    seq_len() or seq_along() instead.

    R/cluster-tools.R:345:12
    R/ncbi-tools.R:162:20
    R/pipeline.R:208:12
    tests/testthat/test-cluster-tools.R:123:17
    tests/testthat/test-download-tools.R:81:17
    ... and 2 more lines

  ✖ fix this R CMD check WARNING: '::' or ':::' imports not
    declared from: ‘ggplot2’ ‘plyr’ Namespaces in Imports field not
    imported from: ‘DBI’ ‘taxize’ All declared Imports should be used.
    Packages in Depends field not imported from: ‘foreach’ ‘methods’
    These packages need to be imported from (in the NAMESPACE file) for
    when this namespace is loaded but not attached. Missing or
    unexported object: ‘doSNOW::registerDoMC’
  ✖ fix this R CMD check NOTE: blstN: no visible binding for
    global variable ‘cmd’ blstN: no visible global function definition
    for ‘read.table’ byLineage : .calc : <anonymous>: no visible
    binding for global variable ‘value’ fltr: no visible global
    function definition for ‘head’ fltr: no visible global function
    definition for ‘tail’ genClstrsObj: no visible global function
    definition for ‘new’ getBestClstrs: no visible global function
    definition for ‘nTaxa’ getBestClstrs: no visible global function
    definition for ‘seqLength’ getClMdLn : <anonymous>: no visible
    global function definition for ‘median’ getMngblIds: no visible
    global function definition for ‘head’ getMngblIds: no visible
    global function definition for ‘tail’ mkBlstDB: no visible binding
    for global variable ‘cmd’ nDscndnts: no visible global function
    definition for ‘%dopar%’ nDscndnts: no visible global function
    definition for ‘foreach’ nDscndnts: no visible binding for global
    variable ‘i’ plotTable: no visible binding for global variable
    ‘clstrnm’ plotTable: no visible binding for global variable
    ‘txidnm’ plotTable: no visible binding for global variable ‘value’
    runStgs: no visible global function definition for ‘is’ safeSrch:
    no visible global function definition for ‘is’ setUp: no visible
    global function definition for ‘packageVersion’ writeClstr: no
    visible global function definition for ‘write.table’ writeSqs: no
    visible binding for global variable ‘clstrs_obj’ writeTax: no
    visible global function definition for ‘write.table’
    summary,ClstrsObj: no visible binding for global variable ‘x’
    Undefined global functions or variables: %dopar% clstrnm clstrs_obj
    cmd foreach head i is median nTaxa new packageVersion read.table
    seqLength tail txidnm value write.table x Consider adding
    importFrom("methods", "is", "new") importFrom("stats", "median")
    importFrom("utils", "head", "packageVersion", "read.table", "tail",
    "write.table") to your NAMESPACE file (and ensure that your
    DESCRIPTION Imports field contains 'methods').
  ✖ fix this R CMD check WARNING: Undocumented data sets:
    ‘blst_rs’ ‘phylt_nds’ ‘sqs’ ‘tdobj’ All user-level objects in a
    package should have documentation entries. See chapter ‘Writing R
    documentation files’ in the ‘Writing R Extensions’ manual.
  ✖ fix this R CMD check NOTE: Warning: no function found
    corresponding to methods exports from ‘phylotaR’ for: ‘show’ A
    namespace must be able to be loaded with just the base namespace
    loaded: otherwise if the namespace gets loaded by a saved object,
    the session will be unable to start. Probably some imports need to
    be declared in the NAMESPACE file.
```

You don't have to address these now.

Seeking reviewers now 🕐

---

Reviewers:  @xx @xx
Due date: xxxx
