### 2018-01-31 #187 @rvosa
@sckott - what would you recommend for a journal to publish a short application note about this package?
