### 2018-01-31 #187 @rvosa
Thanks! Sorry, what's JOSS?

On Wed, Jan 31, 2018 at 8:33 PM, Scott Chamberlain <notifications@github.com
> wrote:

> @rvosa <https://github.com/rvosa> possibly: we have a sort of pairing
> with JOSS and MEE. Also maybe: The R Journal, F1000Research
>
