### 2018-02-06 #187 @sckott
@DomBennett I had a number of questions for you above in my comments. Please respond
