### 2018-02-07 #187 @DomBennett
@sckott Sorry, I thought I could wait for the reviewers and respond to everything wholesale.

**You didn't check the top level MEE box - i've checked that box. You didn't check the box for The manuscript describing the package is no longer than 3000 words - will manuscript length be a problem?**

We hadn't decided at that point which journal to send it to. We're not likely to submit to MEE now. All MEE boxes have been unticked.

**You didn't check the box for has a CRAN and OSI accepted license. I think you do, can you check that box?**

It does now. I added an MIT license.

**You didn't check the box for does not violate the Terms of Service of any service it interacts with. Is that going to be a problem? Can you check on that?**

I've read the terms now and it won't. Box ticked.

**I see you checked the submit to MEE box. Note that we can not make any gaurantee about MEE considering your paper a good fit for them, or them accepting the paper.**

It'd be nice if you did guarantee! But we're not going with MEE....

**Seems like blast on Homebrew (installed via brew install blast) is the same as at the NCBI link you gave at https://www.ncbi.nlm.nih.gov/books/NBK279671/ - Is that right? homebrew is for OSX platforms only**

It's just the link to the general description of how to install BLAST tools across all systems. Not specific to OSX.

**It would indeed be ideal if the pkg didn't have to use system calls to external tools, but there are times when there's no other way.**

Yeah it sucks big time. We're currently working on an approach to avoid using standalone BLAST, in which case a user would not need to install it. The BLAST tools are written in C++, so in theory it should be possible to package the code into the R package. I'm not too au fait on how to do this though. I've written external C scripts into R packages before, but never have I taken someone else's C code entirely. There may be problems with modifying the C++ code to interact with the R. Do you have any ideas on workarounds? Otherwise we could simply give the user the option to send their BLAST jobs to online servers (e.g. NCBI or EBI).

**In terms of style, we prefer snake case but as long as you are consistent within the package that is THE most important thing (i.e, don't use a combination of camelCase, TitleCase and snake_case).**

Yeah I know..... sigh..... Coding style can be real bugbears. I try and use a sort of PEP8 style with a mix of all three (variables snake_case, functions camelCase and classes TitleCase). I can go through and make it all snake_case. I'll probably rechristen some elements where I think the names don't quite work.

**You did not check the box for Does the package conform to rOpenSci packaging guidelines? Can you do that and describe any exceptions**

Not just yet. See above. Also I need to go through your more specific recommendations above as well. Do I need to make these changes before the reviewers look over the package? It's just the pipeline is still being tested and developed and it may be a while before these changes can be made.

**You can put a ropensci review badge in your README**

Dun'd.
