### 2018-02-07 #187 @sckott
> We hadn't decided at that point which journal to send it to. We're not likely to submit to MEE now. All MEE boxes have been unticked.

Okay, sounds good.

> It does now. I added an MIT license.

thanks.

> I've read the terms now and it won't. Box ticked.

thanks.

> We're currently working on an approach to avoid using standalone BLAST ...

Glad you're looking into some options. It's possible reviewers will have some ideas as well.  One idea: if you do shell out, you could use the `sys` package made by Jeroen on our team https://cran.rstudio.com/web/packages/sys/

> I need to go through your more specific recommendations above as well. Do I need to make these changes before the reviewers look over the package?

If you're talking about the `goodpractice` output, no, not now. That's just information for you and reviewers to consider.
