### 2018-02-12 #187 @arendsee
Hi @DomBennett, I'm working on reviewing `phylotaR`, but ran into a few little issues.

 1.  When I run the vignette, I get the following error:

```R
> clstr_id <- getBestClstrs(clstrs_obj, n=1)
Error in nTaxa(clstrs_obj) : could not find function "nTaxa"
```

 2. Fix the BLAST version handling

Currently the BLAST version requirement is hard coded to greater than 2.7.1
(`setup-tools.R:118`). I have 4 issues with this:

 * Hard-coding of values is generally bad.

 * BLAST version 2.7.1 is pretty restrictive. For example, the Ubuntu package
 manager provides BLAST version 2.2.28. Many researchers work in paranoid
 agencies that don't give us admin privileges, so installing the latest and
 greatest is often a pain. Anyway, unless your algorithm depends on new
 features, you should relax the version requirement.

 * The code for testing version is brittle. The version vector is parsed into a 3-tuple (e.g. 2.7.1 ==> `c(2,7,1)`), then tested as below

```R
  # setup-tools.R:118
  tst <- vrsn[1] >= 2 & vrsn[2] >= 7 & vrsn[3] >= 1
```

 However, this would lead to versions such as `2.8.0` being rejected. You can
 fix this by dropping the last comparison:

```R
  # setup-tools.R:118
  tst <- vrsn[1] >= 2 & vrsn[2] >= 7
```

Also, in DESCRIPTION, the  `SystemRequirements` specifies a BLAST version of `>2.7.1`, but your actual version checking function requires `>=2.7.1`.

 3. Fix the warnings and notes in `check()`

 When I run `devtools::check()` I get 5 warnings, and 3 notes. The issues all look pretty normal (easy to fix). I don't see anything too serious here. Below I have outlined the main issues and how to fix them (I might have missed one or two).

 * Remove the `warnings_are_errors: false` clause from your `.travis.yml`
 file. CRAN will generally not accept a package if any warnings are raised by
 CHECK. So you probably should not ignore warnings in your continuous
 integration.

 * DESCRIPTION says that your license is GPL-2, but the actual LICENSE file is
 MIT. You need to choose one of them. `devtools` makes license handling pretty
 easy. Just delete your current LICENSE file and the `License` entry in your
 DESCRIPTION file, and then call either `devtools::use_mit_license()` or
 `devtools::use_gpl3_license()`. This will create the license and update the
 DESCRIPTION accordingly.

 * You need to add `ggplot2` and `plyr` to your `Imports` list in DESCRIPTION

 * You need to add the statements:

```R
#' @importFrom methods is new
#' @importFrom stats median
#' @importFrom utils heads packageVersion read.table tail write.table
```

 Somewhere in your documentation (I usually put these lines in the same place
 where I put package level documentation).

 * Describe each parameter in the documentation

Cheers!
