### 2018-02-13 #187 @DomBennett
Thanks for the detailed and early feedback. Dealing with your specific issues:

1.  nTaxa
The functions and classes associated with the clusters object are still a little open-ended at this stage, hence why their names may change. The equivalent to `nTaxa` is now `getClNTx` - get n taxa by cluster. We're still trying to determine the sorts of ways users will want to interrogate the clusters, and the best sort of language to use when naming functions.

I've updated the vignette. But the code is likely to change in the near future and will need changed again.

2. BLAST
OK. Thanks for the tips. Fixed and updated. Now using v 2.0 and up. How would you test the versioning without hard coding the values?

3. Check warnings

* travis warnings - removed from yaml
* licensing updated to MIT
* ggplot2 and plyr - added
* statements added to phylotaR.R
* parameters are being worked and should be updated soon!
