### 2018-02-14 #187 @arendsee
> OK. Thanks for the tips. Fixed and updated. Now using v 2.0 and up. How would you test the versioning without hard coding the values?

Hmm, good point. @sckott Any thoughts on best practices for confirming the version of an outside tool (BLAST, in this case)?

I can tell you what I do, though I am not entirely happy with my approach. I simply don't check the version of my tools, instead I just print the version output into the log (e.g. system call to `blastp -version`).

If you want to check the version, you could write a dedicated function to extract a version string from a version statement (maybe `get_version` or something), and then write a dedicated version comparison function (maybe `cmp_version`).

I don't really have a good general solution. But in your case, I would recommend not checking the version. There are differences between the BLAST versions, but I don't _think_ any of the BLAST+ versions anyone will actually use (i.e. >=2.2) will change the results. Perhaps it would be worthwhile to test several versions of BLAST on Travis.

Thoughts? @sckott @naupaka?
