### 2018-02-14 #187 @sckott
I don't know off hand about best practices for checking versions of cli tools. Thoughts @jeroen @richfitz ?

> Perhaps it would be worthwhile to test several versions of BLAST on Travis.

this seems like a really good idea - and then any changes that affect certain BLAST versions and not others will become immediately apparent
