### 2018-02-14 #187 @arendsee
@naupaka I agree on pulling BLAST into R being a bad idea. Unless NCBI was interested in doing the work and maintaining the code.

As for the legacy BLAST versus BLAST+, the names of the executable have all changed. For example, `blastp` is not defined in legacy, they used `blastall` with some protein-vs-protein argument, I think. So if someone tried to use legacy BLAST the commands required simply would not be found.
