### 2018-02-15 #187 @DomBennett
OK. So... actionable points from this discussion:
- Don't test for BLAST version
  - Print version in log
  - But do ensure it's not legacy BLAST
- Write tests that run across different BLAST versions for travis
- Petition NCBI to support BLAST in R
