### 2018-02-15 #187 @arendsee
@DomBennett I probably got a bit beyond the scope of this review in my earlier comments. I'd say the only actionable point in regards to `phylotaR` is the first bullet.

Tests in Travis against multiple versions of BLAST would be cool, but as far as I am concerned, you don't need to do that now.
