### 2018-03-11 #187 @arendsee
This is solid work and I recommend it be admitted to rOpenSci after a few
issues are addressed. The documentation looks good, program seems to work as
intended, the interface to the output objects is clear, and the results are
neatly cached and logged. This is a good example of how a clean R pipeline
should look.

My one high-level comment is that this package could benefit greatly by reusing
some existing R packages: `taxize` and `taxizedb` can handle NCBI taxonomy IDs;
bioconductor's `Biostrings` can be used for reading, writing, and translating
sequence; possibly `drake` could be used as a build engine.

# Package Review

Please check off boxes as applicable, and elaborate in comments below. Your
review is not limited to these topics, as described in the reviewer guide

 * [x] As the reviewer I confirm that there are no conflicts of interest for me
 to review this work (If you are unsure whether you are in conflict, please
 speak to your editor before starting your review).

 * [x] A statement of need clearly stating problems the software is designed to
 solve and its target audience in README

 * [x] Installation instructions: for the development version of package and any
 non-standard dependencies in README

 * [x] Vignette(s) demonstrating major functionality that runs successfully
 locally

 * [x] Function Documentation: for all exported functions in R help

 * [ ] Examples for all exported functions in R Help that run successfully locally

**No examples given**

 * [ ] Community guidelines including contribution guidelines in the README or
 CONTRIBUTING, and DESCRIPTION with URL, BugReports and Maintainer (which may
 be autogenerated via Authors@R).

**Missing CONTRIBUTING**

 * [x] Installation: Installation succeeds as documented.

 * [x] Functionality: Any functional claims of the software been confirmed.

 * [x] Performance: Any performance claims of the software been confirmed.

There aren't really any performance claims made. However, the vignette ran
pretty quickly.

 * [ ] Automated tests: Unit tests cover essential functions of the package and
 a reasonable range of inputs and conditions.

Testing coverage is at 49%. I'd recommend spending some time nailing all the
functions down with direct comparisons to their expected results. It is also
worthwhile to check all the corner cases.

 * [x] All tests pass on the local machine.

 * [ ] Packaging guidelines: The package conforms to the rOpenSci packaging
 guidelines

 **- missing examples**

## Final approval (post-review)

 * [ ] The author has responded to my review and made changes to my satisfaction.
 I recommend approving this package.

## Review Comments

### Contributing

`rOpenSci` recommends you add a CONTRIBUTING file to your repo.

 * [ ] I would recommend a version formatted as `<major>.<minor>.<patch>` (e.g.
 `0.1.0`), [see here](http://r-pkgs.had.co.nz/description.html#version). This
 is by far the most common convention in the R world (and in the open source
 world in general).

 * [ ] The `Date` field should not be used

 * [ ] For the `URL` and `BugReports` entries, it would be better if these
 pointed to the repository that holds the source code (see `Repos issues`
 section below).

 * [ ] The `License` slot says GPL-2 but your actual LICENSE is MIT, this needs
 to be resolved.

### Style

Skimming the code, everything looks tidy and well organized. Functions are
partitioned between files in an understandable way and every function is
documented (see `CHECK` for exceptions to this). I have just a few little

 * [ ] Your line wrapping is a little irregular, for example, in the description
 of the `run` function. Some function documentation is wrapped at 80 lines,
 some at 70, and some at 50 (e.g. `runTaxise`). This is not a big deal, but it
 is a little awkward to read.

 * [ ] There is an extra newline at the beginning of some, but not all, of your files.

The documentation is pretty solid. Just a few comments:

 * [ ] There are a few miss-spellings in the documentation (e.g.
 `cluster-tools.R:340`). You should run `devtools::spell_check()` and fix all
 mistakes (don't worry about the false positives).

 * [ ] Since R is dynamically typed, it is often useful to specify the expected
 type for each parameter and the return type in the documentation. The input
 types are covered reasonably well in the `@param` fields, but a lot of
 functions are missing the `@return` field. I strongly recommend adding
 a `return` field to every function, even if it returns nothing.

 * [ ] According to the rOpenSci guidelines, each function should have examples
 in the documentation. You might consider creating example data for your
 package, that can be loaded and used in the examples. For example, you could
 wrap all the data from your vignette example into a list, and then load it
 with `data(aotus)`. The `devtools::use_data` function makes this pretty easy
 to do. This approach circumvents the need to rerun a whole analysis in each
 example.

 * [ ] Some of the function and class names are a little cryptic. I would prefer
 longer, more descriptive and memorable names, for example `ClusterRecord`
 rather than `ClRcrd` (which I might interpret as `CladeRecord`, among other
 possibilities).

 * [ ] Snakecase is usually preferred in R for variable names. I'm fine with
 class names being camel case (that is how bioconductor does it).

### Repo issues

 The `DomBennett/phylotaR` repo is forked from the `AntonelliLab/phylotaR`
 repo. According to DESCRIPTION, issues should be logged at
 `AntonelliLab/phylotaR`. I don't think this is a good idea -- issues should
 not be separated from their source. There are at least two solutions:

 1. Allow issues to be posted on the `DomBennett/phylotaR` repo. By default you
    cannot post issues on a forked repo. However, this default can be
    [overridden](https://softwareengineering.stackexchange.com/questions/179468).

 2. Merge the fork back into `AntonelliLab/phylotaR`.

Overall the README looks good. Here are a few suggestions:

  * [ ] Delete the comment, "Please note: this package is in its development and
  is not fully tested.". Before this goes into rOpenSci, you should be
  reasonably confident that the program works. There are always bugs, but we
  everyone expects this, so no special mention is needed.

  * [ ] Specify the minimum required BLAST+ version.

  * [ ] In the `Pipeline` section, I might break the first paragraph into
  a numbered list of steps; then elaborate a *little* more. Specifically, you
  should say how you identify orthologous clusters with BLAST. Then maybe
  another sentence of explanation on `cluster^2`. Overall, the `Pipeline`
  section is great and I like your figure.

  * [ ] The naming you use in the `Pipeline` section is a little confusing.
  I recommend using the same names you use in your code. Rather than saying
  `taxise`, `download`, `cluster`, and `cluster^2`; say `runClusters`,
  `runClusters2`, `runDownload`, and `runTaxise`. This is especially confusing
  for `taxise`, where I think you are referring to `runTaxise` which is
  performing a function that is very similar to what might be done in the
  `taxize` package (which mixed me up at first).

### Vignette

The introductory vignette is well written and clear. There is a reference to an
advanced vignette, but this vignette is not yet written. Here are my
recommendations:

  * [ ] In the introductory vignette, add a section discussing how long the
  program will take to run.

  * [ ] Either open a github issue or a TODO list entry for completion of the
  advanced vignette. As far as my review is concerned, I am OK with this
  submission lacking the vignette.

### Running the code

Here I will walk through the vignette and add suggestions in checkboxes.

```R
library(phylotaR)
wd <- 'aotus'
ncbi_dr <- '/usr/bin'
txid <- 9504
setUp(wd=wd, txid=txid, ncbi_dr=ncbi_dr, v=TRUE)
#R> -----------------------------------------------
#R> phylotaR: Implementation of PhyLoTa in R * [v0.1]
#R> -----------------------------------------------
#R> Checking for valid NCBI BLAST+ Tools ...
#R> Found: [/usr/bin/makeblastdb]
#R> Found: [/usr/bin/blastn]
#R> Setting up pipeline with the following parameters:
#R> . blstn      [/usr/bin/blastn]
#R> . btchsz     [300]
#R> . date       [2018-03-11]
#R> . mdlthrs    [3000]
#R> . mkblstdb   [/usr/bin/makeblastdb]
#R> . mncvrg     [51]
#R> . mnsql      [250]
#R> . mxevl      [1e-10]
#R> . mxnds      [1e+05]
#R> . mxrtry     [100]
#R> . mxsql      [2000]
#R> . mxsqs      [50000]
#R> . ncps       [1]
#R> . txid       [9504]
#R> . v          [TRUE]
#R> . wd         [aotus]
#R> -----------------------------------------------
```

This output summary is concise and informative and is recorded in a log file.

 * [ ] The log should include the NCBI BLAST version.

 * [ ] It also might be nice to include the `sessionInfo()` results in the log.
 Or you could write it into a separate file, if it clutters the log too much.

Next we run the program:

```R
run(wd=wd)
```

This prints results to the screen and the log. There are quite a few messages
like this:

```
. Processing [id 9504] child [id 231953]
. . Generating subtree clusters for [id 231953(species)]
. . . . [0 sqs] -- too few sequences, cannot make clusters
```

 * [ ] You might want a little "Don't Panic" comment saying that this is OK (it
 is, right?).

Next we explore the results (still following the vignette):

```R
all_clustrs <- read_phylota(wd)
cids <- all_clustrs@cids
n_taxa <- get_ntaxa(phylota=all_clustrs, cid=cids)
```

I really liked the care that was put into making informative print statements.

 * [ ] The `show` and `print` generics are a little inconsistent. I recommend
 stepping through all your classes and ensuring that explicit calls to `print`,
 `show`, and the other generics are doing exactly what you want. For example:

```R
all_clustrs@sqs
# you probably want `cat` not `print` in `all_classes.R:301`
print(all_clustrs@sqs)
show(all_clustrs@sqs)
```

 * [ ] It might be worthwhile to print a one line description of each slot when
 printing an object. This would let the user interact with the objects without
 having to refer back to the docs. I'm not sure about this, though, since it
 might just be too cluttered.

Moving on:

``` R
keep <- cids[n_taxa > 6]
selected <- drop_cls(phylota=all_clustrs, cid=keep)
smmry <- summary(selected)
```

`summary(selected)` makes a table with a row for each table. I really like this
use of `summary`, good information, but two points:

 * [ ] You should document somewhere what MAD means. Also Type and Seed.

 * [ ] In finding the most common words in the description and gene name, you
 might want to filter out common words. You might want to look into the `tm`
 library.

Continuing on to the last part of the vignette:

```R
# choose best sequence per species
reduced <- drop_by_rank(phylota=selected, rnk='species', n=1)
# get txids at the species level for each sequence
txids <- get_txids(phylota=reduced, cid=cid, rnk='species')
# look up name for txids
scientific_names <- get_tx_slot(phylota=reduced, txid=txids, slt_nm='scnm')
# clean the names
scientific_names <- gsub('\\.', '', scientific_names)
scientific_names <- gsub('\\s+', '_', scientific_names)
# look up sequence IDs for our chosen cluster
sids <- reduced@cls[[cid]]@sids
# write out
write_sqs(phylota=reduced, sid=sids, outfile='cytb.fasta',
          sq_nm=scientific_names)
```

 * [ ] I think `phylotaR` would benefit from reusing some existing libraries.
 `taxize` and `taxizedb` offer a lot of functions for mapping between IDs and
 scientific names. Also, for writing and reading FASTA sequencees, you might
 consider using the bioconductor Biostrings library.

 * [ ] FASTA files are normally formatted with on spaces between entries and with
 the sequences wrapped (usually at 60 or 80 characters). Again, you might want
 to use the Biostrings library.
