### 2018-03-26 #187 @naupaka
I'm just about done with my review. Have had to let it run overnight a couple times and hit some snags. I'm hoping to have the review done by tomorrow late morning if all goes well on this current query.
