### 2018-03-26 #187 @naupaka
- [X] As the reviewer I confirm that there are no conflicts of interest for me to review this work.

### Overall

I think this is a great and super useful package that I think is a good fit with ROpenSci. I am excited for this package to get finished up; I am already thinking up half a dozen ways to use it in my own research. All of that said, I had some problems getting it to run on my machine (even though it built and installed fine), and the documentation and examples are still being worked on. I'd be happy to take another look once it's further along, and like I said, I think this will be a really useful package for a whole lot of people, so I am grateful to the authors for putting it together. I agree with @arendsee's comments about considering more use of existing packages, especially `taxize` and `Biostrings` to simplify the code and reduce duplication of effort.

   There is a link to a website detailing this info, but it is not covered in the README in the repo.

   Installation was straighforward, despite the external dependency.

  There is a nice introductory vignette, but I had some trouble running through the examples (details below).

   The Description and Examples are still missing on a number of functions, e.g. `runTaxize()`.

   A good number of functions are still lacking examples.

   No info on contributing yet, either in README or in a CONTRIBUTING file.

   I cannot run through the pipeline suggessfully - I've tried with four different Taxa IDs and it fails each time on the `makeblastdb` step.

   No claims made, but I think it would be great to include some ballpark info on how long a given run is going to take. If you don't `tail -f` the log file, it's hard to tell what's going on. The output in the R environment is very sparse.

- [ ] Automated tests: Unit tests cover essential functions of the package

   Accourding to `covr()`, about half of the code is covered by tests.

- [ ] Packaging guidelines: The package conforms to the rOpenSci packaging guidelines

   Getting there!

Estimated hours spent reviewing: 8 (not including overnight runs through the pipeline)

---

- Does the code comply with general principles in the Mozilla reviewing guide?
   - I think so.
- Does the package comply with the ROpenSci packaging guide?
   - Getting there!
- Are there improvements that could be made to the code style?
   - I know it would be a good amount of work, but I think the function names could be standardized with a prefix, since some of them are very common terms and there could be confusion if they mask other functions of the same name. I also think the naming style in general is very condensed, and it wouldn't hurt to expand the names of functions and other objects a little in the codebase to aid in readability and facilitate contributions from others.
- Is there code duplication in the package that should be reduced?
   - Not as far as I can tell.
- Are there user interface improvements that could be made?
   - I think there could be a bit more in terms of output for the user when long-running functions are working. The log file on disk is great, and the disk file/folder organization is very neat and well-organized, but it's hard to tell what's going on in R. Since running the standard pipeline can take hours to tens of hours (in my experience), it would be nice to know all is still progressing.
- Are there performance improvements that could be made?
   - Don't think so. I imagine the rate-limiting step is the speed at which you can hit the NCBI API.
- Is the documentation (installation instructions/vignettes/examples/demos) clear and sufficient?
   - They are still in progress, and could use a bit more filling out (but I see this is still a work in progress).

We encourage you to use automated tools to facilitate your reviewing. These include:

- Running devtools::check() and devtools::test() on the package to find any errors that may be missed on the author's system.
   - No errors with either of these, but `R CMD check` returns a number of WARNINGS and NOTES.
- Using the goodpractice package (goodpractice::gp()) to identify likely sources of errors and style issues.
   - Already pasted above by @sckott
- Using devtools::spell_check() to find spelling errors in DESCRIPTION and manual files, and RStudio's spellchecker to find spelling errors in vignettes and README files.
   - Since these are still in progress, I imagine it would be more useful to check later.

#### Other Comments

```r
> devtools::session_info()

Session info -----------------------------------------------------------------------------------------------------------
 system   x86_64, darwin17.3.0
 ui       RStudio (1.2.472)
 collate  en_US.UTF-8
 tz       America/Los_Angeles

Packages ---------------------------------------------------------------------------------------------------------------
 package   * version date       source
 backports   1.1.2   2017-12-13 CRAN (R 3.4.3)
 base      * 3.4.4   2018-03-15 local
 base64enc   0.1-3   2015-07-28 CRAN (R 3.4.0)
 compiler    3.4.4   2018-03-15 local
 datasets  * 3.4.4   2018-03-15 local
 devtools    1.13.5  2018-02-18 CRAN (R 3.4.3)
 digest      0.6.15  2018-01-28 CRAN (R 3.4.3)
 evaluate    0.10.1  2017-06-24 CRAN (R 3.4.2)
 graphics  * 3.4.4   2018-03-15 local
 grDevices * 3.4.4   2018-03-15 local
 htmltools   0.3.6   2017-04-28 CRAN (R 3.4.0)
 jsonlite    1.5     2017-06-01 CRAN (R 3.4.0)
 knitr       1.20    2018-02-20 CRAN (R 3.4.3)
 magrittr    1.5     2014-11-22 CRAN (R 3.4.0)
 memoise     1.1.0   2017-04-21 CRAN (R 3.4.0)
 methods   * 3.4.4   2018-03-15 local
 phylotaR  * 0.1     2018-03-26 Github (DomBennett/phylotaR@2541b03)
 Rcpp        0.12.16 2018-03-13 CRAN (R 3.4.3)
 rmarkdown   1.9     2018-03-01 CRAN (R 3.4.3)
 rprojroot   1.3-2   2018-01-03 CRAN (R 3.4.3)
 stats     * 3.4.4   2018-03-15 local
 stringi     1.1.7   2018-03-12 cran (@1.1.7)
 stringr     1.3.0   2018-02-19 CRAN (R 3.4.3)
 sys         1.5     2017-10-10 cran (@1.5)
 tools       3.4.4   2018-03-15 local
 treeman     1.1.1   2017-06-27 cran (@1.1.1)
 utils     * 3.4.4   2018-03-15 local
 withr       2.1.2   2018-03-15 CRAN (R 3.4.3)
 yaml        2.1.18  2018-03-08 CRAN (R 3.4.3)

> sessionInfo()

Platform: x86_64-apple-darwin17.3.0 (64-bit)
Running under: macOS High Sierra 10.13.3

BLAS: /System/Library/Frameworks/Accelerate.framework/Versions/A/Frameworks/vecLib.framework/Versions/A/libBLAS.dylib
LAPACK: /System/Library/Frameworks/Accelerate.framework/Versions/A/Frameworks/vecLib.framework/Versions/A/libLAPACK.dylib

[1] en_US.UTF-8/en_US.UTF-8/en_US.UTF-8/C/en_US.UTF-8/en_US.UTF-8

[1] phylotaR_0.1

 [1] Rcpp_0.12.16    treeman_1.1.1   digest_0.6.15   withr_2.1.2     rprojroot_1.3-2 jsonlite_1.5    sys_1.5
 [8] backports_1.1.2 magrittr_1.5    evaluate_0.10.1 stringi_1.1.7   rstudioapi_0.7  rmarkdown_1.9   devtools_1.13.5
[15] tools_3.4.4     stringr_1.3.0   yaml_2.1.18     compiler_3.4.4  base64enc_0.1-3 memoise_1.1.0   htmltools_0.3.6
[22] knitr_1.20
```

#### Trying to run workflow pipeline on various taxa, encountering troubles

```r
> devtools::install_github(repo='DomBennett/phylotaR', build_vignettes=TRUE)
# https://github.com/DomBennett/phylotaR/commit/2541b030158183b8b1d3c1f44c0ea26e96ab11f3
library("phylotaR")
```

```r
wd <- "~/Desktop/phylota_review/primates"
ncbi_dr <- "/usr/local/bin/"
txid <- 9443 #primates
setUp(wd=wd, txid=txid, ncbi_dr=ncbi_dr)
run(wd=wd)
```

Hung when trying to download chimpanzees, no progress in log file for 20+ minutes.

> . . Working on child [id 9598]
> . . + direct ...

Trying again with a smaller clade: Mycosphaerella (fungal genus) -- 41254

```r
wd <- "~/Desktop/phylota_review/fungi"
ncbi_dr <- "."
txid <- 41254 #Mycosphaerella
setUp(wd=wd, txid=txid, ncbi_dr=ncbi_dr)
```

`ncbi_dr <- "."` did not work as recommended in the vignette despite the programs being in my `PATH`.

```
Error in setUpNcbiTools(d = ncbi_dr, v = v, wd = wd) : Unable to find correct versions of NCBI BLAST+ tools
3. stop(msg)
2. setUpNcbiTools(d = ncbi_dr, v = v, wd = wd)
1. setUp(wd = wd, txid = txid, ncbi_dr = ncbi_dr)
```

Works fine with absolute path to directory.

```r
wd <- "~/Desktop/phylota_review/fungi"
ncbi_dr <- "/usr/local/bin"
txid <- 41254 #Mycosphaerella (large genus)
setUp(wd=wd, txid=txid, ncbi_dr=ncbi_dr)
run(wd=wd)
```

I kept getting the following error, even on different initial query taxa IDs.

```
Error in runStgs(wd = wd, frm = 1, to = nstages, stgs_msg = stgs_msg) : Unexpected Error in error(ps = ps, paste0("makeblastdb failed to run. Check BLAST log files.")) : Error: makeblastdb failed to run. Check BLAST log files. Occurred [2018-03-25 20:45:13] Contact package maintainer for help.
3. stop(msg)
2. runStgs(wd = wd, frm = 1, to = nstages, stgs_msg = stgs_msg)
1. run(wd = wd)
```

```
---------------------------------------------
Starting stage CLUSTER: [2018-03-25 20:45:12]
---------------------------------------------
Working on [id 41254]
. Generating subtree clusters for [id 41254(genus)]
. Generating direct clusters for [id 41254(genus)]
. . [0 sqs]
. . . Too few sequences, cannot make clusters
. BLASTing [1079 sqs] ....
. . Running makeblastdb
Error: makeblastdb failed to run. Check BLAST log files.
Unexpected Error in error(ps = ps, paste0("makeblastdb failed to run. Check BLAST log files.")) :
  Error: makeblastdb failed to run. Check BLAST log files.

Occurred [2018-03-25 20:45:13]
Contact package maintainer for help.
```

Didn't work because not enough sequence diversity. Trying again with a higher level group.

```r
wd <- "~/Desktop/phylota_review/capnodiales"
ncbi_dr <- "/usr/local/bin"
txid <- 134362 # Capnodiales (fungal family)
setUp(wd=wd, txid=txid, ncbi_dr=ncbi_dr)
run(wd=wd)
```

```
... Taxise
... Download
... Cluster
Error in runStgs(wd = wd, frm = 1, to = nstages, stgs_msg = stgs_msg) :
  Unexpected Error in error(ps = ps, paste0("makeblastdb failed to run. Check BLAST log files.")) :
  Error: makeblastdb failed to run. Check BLAST log files.

Occurred [2018-03-25 23:30:10]
Contact package maintainer for help.
```

Here's the output from the blast log file.

```
# blast log file
$ cat blast/taxon-1047167-typ-subtree-db.log

BLAST options error: File ~/Desktop/phylota_review/capnodiales/blast/taxon-1047167-typ-subtree-db.fa does not exist
w/capnodiales/blast/taxon-1047167-typ-subtree-db.fa
New DB title:  ~/Desktop/phylota_review/capnodiales/blast/taxon-1047167-typ-subtree-db.fa
Sequence type: Nucleotide
Keep MBits: T
Maximum file size: 1000000000B
```

Some thoughts on this process:

* It's unclear what makes a suitable clade (during the search process).
* There is helpful info on how many out of how many parental clades have been downloaded, but since it's unclear how many children for each there might be and how many sequences in each of those, it is hard to tell how long it will all take. Is there any way to estimate and report this before starting?
* For (e.g.) genera with many species in them, it is unclear why the dots in the logfile keep increasing. If all the species are sister to one another, shouldn't they not be indented under one another?

```r
restart(wd)
```

Does not restart wthout restarting the R session.

#### Trying to run demo in aotus.R:

Also using a absolute instead of a relative path for the working directory here, in case that is screwing up `makeblastdb`, but it still didn't work, as before.

```
> # RUN PIPELINE
> txid <- 9504
> setUp(wd="/Users/naupaka/Desktop/phylota_review/aotus", txid=txid, ncbi_dr=ncbi_dr, v=TRUE)
-----------------------------------------------
phylotaR: Implementation of PhyLoTa in R [v0.1]
-----------------------------------------------
Checking for valid NCBI BLAST+ Tools ...
Found: [/usr/local/bin/makeblastdb]
Found: [/usr/local/bin/blastn]
Setting up pipeline with the following parameters:
. blstn      [/usr/local/bin/blastn]
. btchsz     [300]
. date       [2018-03-26]
. mdlthrs    [3000]
. mkblstdb   [/usr/local/bin/makeblastdb]
. mncvrg     [51]
. mnsql      [250]
. mxevl      [1e-10]
. mxnds      [1e+05]
. mxrtry     [100]
. mxsql      [2000]
. mxsqs      [50000]
. ncps       [1]
. txid       [9504]
. v          [TRUE]
. wd         [/Users/naupaka/Desktop/phylota_review/aotus]
-----------------------------------------------
> run(wd=wd)
... Taxise
... Download
... Cluster
Error in runStgs(wd = wd, frm = 1, to = nstages, stgs_msg = stgs_msg) :
  Unexpected Error in error(ps = ps, paste0("makeblastdb failed to run. Check BLAST log files.")) :
  Error: makeblastdb failed to run. Check BLAST log files.

Occurred [2018-03-26 06:37:35]
Contact package maintainer for help.
```

#### Output of R CMD check

```sh
$ R CMD check .
* using log directory ‘/Users/naupaka/Desktop/phylota_review/phylotaR/..Rcheck’
* using R version 3.4.4 (2018-03-15)
* using platform: x86_64-apple-darwin17.3.0 (64-bit)
* using session charset: UTF-8
* checking for file ‘./DESCRIPTION’ ... OK
* checking extension type ... Package
* this is package ‘phylotaR’ version ‘0.1’
* checking package namespace information ... OK
* checking package dependencies ... OK
* checking if this is a source package ... OK
* checking if there is a namespace ... OK
* checking for executable files ... OK
* checking for hidden files and directories ... NOTE
Found the following hidden files and directories:
  .gitignore
  .travis.yml
  ..Rcheck
  .git
These were most likely included in error. See section ‘Package
structure’ in the ‘Writing R Extensions’ manual.
* checking for portable file names ... OK
* checking for sufficient/correct file permissions ... OK
* checking whether package ‘phylotaR’ can be installed ... OK
* checking installed package size ... OK
* checking package directory ... OK
* checking DESCRIPTION meta-information ... NOTE
Checking should be performed on sources prepared by ‘R CMD build’.
* checking top-level files ... OK
* checking for left-over files ... OK
* checking index information ... OK
* checking package subdirectories ... WARNING
Found the following directory with the name of a check directory:
  ./..Rcheck
Most likely, these were included erroneously.
* checking R files for non-ASCII characters ... OK
* checking R files for syntax errors ... OK
* checking whether the package can be loaded ... OK
* checking whether the package can be loaded with stated dependencies ... OK
* checking whether the package can be unloaded cleanly ... OK
* checking whether the namespace can be loaded with stated dependencies ... OK
* checking whether the namespace can be unloaded cleanly ... OK
* checking loading without being on the library search path ... OK
* checking dependencies in R code ... OK
* checking S3 generic/method consistency ... OK
* checking replacement functions ... OK
* checking foreign function calls ... OK
* checking R code for possible problems ... OK
* checking Rd files ... OK
* checking Rd metadata ... OK
* checking Rd cross-references ... OK
* checking for missing documentation entries ... OK
* checking for code/documentation mismatches ... OK
* checking Rd \usage sections ... OK
* checking Rd contents ... OK
* checking for unstated dependencies in examples ... OK
* checking contents of ‘data’ directory ... OK
* checking data for non-ASCII characters ... OK
* checking data for ASCII and uncompressed saves ... OK
* checking files in ‘vignettes’ ... WARNING
Files in the 'vignettes' directory but no files in 'inst/doc':
  ‘phylotaR-advanced.Rmd’, ‘phylotaR-advanced.html’, ‘phylotaR.R’,
    ‘phylotaR.Rmd’, ‘phylotaR.html’
* checking examples ... OK
* checking for unstated dependencies in ‘tests’ ... OK
* checking tests ...
  Running ‘test-all.R’
 OK
* checking for unstated dependencies in vignettes ... OK
* checking package vignettes in ‘inst/doc’ ... WARNING
Package vignettes without corresponding PDF/HTML:
   ‘phylotaR-advanced.Rmd’
   ‘phylotaR.Rmd’

* checking running R code from vignettes ...
   ‘phylotaR-advanced.Rmd’ using ‘UTF-8’ ... OK
   ‘phylotaR.Rmd’ using ‘UTF-8’ ... OK
 NONE
* checking re-building of vignette outputs ...Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
Warning in file.copy(pkgdir, vd2, recursive = TRUE) : too deep nesting
[…]
