### 2018-03-27 #187 @DomBennett
Thanks very much @arendsee and @naupaka for your detailed reviews!

I feel I may have already addressed some of the run issues raised by @naupaka, as I encountered some of them over the past few weeks myself. The hanging is particularly annoying and seems to be due to Entrez, I've added a timeout so that it will retry after one hour. Also, I wonder if the BLAST errors might be due to using Windows, I've not tested the pipeline at all on that OS. Lots of things to be getting on with!

In terms of next steps @sckott , do I just update the package and report what I did here?
