### 2018-04-04 #187 @DomBennett
Hi @sckott,

I've copied your format and created [issues](https://github.com/AntonelliLab/phylotaR/issues). Also, FYI we have now submitted a manuscript describing phylotaR to MDPI Life, preprint available [here](https://www.preprints.org/manuscript/201804.0047/v1).

Is there a deadline for responding to the reviewers? I was thinking of taking a break and coming back to phylotaR with a little more distance.
