### 2018-04-05 #187 @sckott
Not really a deadline. If you'll put it off for a while, we have a [holding label](https://github.com/ropensci/onboarding/issues?q=is%3Aissue+is%3Aopen+label%3Aholding) that we'll use, and then we'll know to not ask you about progress until you're ready to come back to it
