### 2018-06-16 #187 @DomBennett
Hi @sckott, @arendsee and @naupaka,

I’ve made the following major changes as per your comments and reviews:

- **Restyled**:  All functions and variables are snake case with the exception of the S4 classes that remain camelcase. I’ve also updated the names to make things a little more explicit (e.g. ClstrRec rather than ClRec).
- **Tested**: I now have [89% coverage](https://coveralls.io/github/AntonelliLab/phylotaR?branch=master)
- **Examples**: All ‘public’, exported functions and data have examples.
- **Pkgdown**: I have created a [website](https://antonellilab.github.io/phylotaR/).
- **Documentation**: all functions have return parameters, I have checked spelling with spell_check.
- **Timings**: The demonstration phylotaR runs now have the timings of each stage on the README.md
- **Contributing**: There is now a CONTRIBUTING.md
- **Logging**: There is now more salient information printed to screen, blast versions and session info are also recorded in a user’s wd.
- **Windows**: The pipeline has been run on a windows machine
- **Makeblastdb error**: I think this was due to a ‘~’ in @naupaka’s filepath. I now halt the pipeline if a user’s does this.

Things I have not done:

- **Additional packages**: I have not incorporated taxize, taxizeDB, tm or Biostrings into the package. In the case of the latter, I feel it best to avoid Bioconductor packages as they are a little more tricky to install and I would like to keep phylotaR light, even though it requires BLAST+. In the case of the others, I see no immediate reason to make use of these additional packages while I currently have no need for their additional capabilities. I have fixed the 80 column width issue of the user generated .fasta output. In the future, as new features are requested (e.g. alternative taxonomies), then they should be incorporated. Right now though, I feel I should follow the doctrine of if it ain’t broke, why fix it?
- **Advanced vignette**: I have not added this as a lot of the additional information can be found in the recently [published article](https://doi.org/10.3390/life8020020) on the pipeline. In the future I can imagine a new vignette containing more detailed information on new features with more complex options. This is likely to occur since I have been contacted a fair number of times with requests for new options and features.

If any of you feel that these last two points should be acted upon in some form, then I will be happy to look at these again.

Many thanks once more for taking the time to review the package!

Best wishes,
Dom
