### 2018-06-16 #187 @naupaka
Sounds great @DomBennett!

@sckott shall we have another go through of it with the new version? What's the ROpenSci SOP on review after revisions?
