### 2018-06-20 #187 @sckott
thanks @DomBennett !

@naupaka @arendsee

If you're completely happy with changes to the package and responses from the maintainer then just say so. If not, please do comment here with any additional comments/suggestions/etc.
