### 2018-06-20 #187 @arendsee
@DomBennett @sckott Looks good to me.

I would recommend *eventually* refactoring with `taxize` et al, but I wouldn't consider it a necessity.

Good job!
