### 2018-06-20 #187 @sckott
@naupaka do you want to make further comments, or are you happy?
