### 2018-06-20 #187 @naupaka
I'd like to put the new version through its paces and haven't had the time to sit down and do it. Will get back to you in the next couple days if that's ok?
