### 2018-07-08 #187 @naupaka
@DomBennett Just ran through a couple testing workflows and everything works smoothly for me! Thanks for this very useful tool.
