### 2018-07-09 #187 @sckott
@DomBennett we're nearly there. just a few comments:

On running the example from the vignette, I get an error:

```r
---------------------------------------------------
Running pipeline on [unix] at [2018-07-09 10:22:31]
---------------------------------------------------
Running stages: taxise, download, cluster, cluster2
--------------------------------------------
Starting stage TAXISE: [2018-07-09 10:22:31]
--------------------------------------------
Searching taxonomic IDs ...
Downloading taxonomic records ...
. [1-21]
Generating taxonomic dictionary ...
---------------------------------------------
Completed stage TAXISE: [2018-07-09 10:22:33]
---------------------------------------------
----------------------------------------------
Starting stage DOWNLOAD: [2018-07-09 10:22:33]
----------------------------------------------
Identifying suitable clades ...
Identified [1] suitable clades.
Downloading hierarchically ...
Working on parent [id 9504]: [1/1] ...
. + whole subtree ...
. . Getting [2700 sqs] ...
. . . [1-300]
Unexpected Error : NCBI is limiting the size of your request. Consider reducing btchsz.

Occurred [2018-07-09 10:22:58]
Contact package maintainer for help.
Error in stages_run(wd = wd, frm = 1, to = nstages, stgs_msg = stgs_msg) :
  Unexpected Error : NCBI is limiting the size of your request. Consider reducing btchsz.

Occurred [2018-07-09 10:22:58]
Contact package maintainer for help.
```

It looks like I need to fiddle with `btchsz` but I'm not sure where to do that. Would it make sense to say what function has that parameter? It looks like it's the function `parameters()`?

running goodpractice again: i'd encourage you to avoid `sapply` where possible and avoid 1:length(...), 1:nrow(...), 1:ncol(...), 1:NROW(...) and 1:NCOL(...) type expressions where possible,  but not there aren't required before approval.
