### 2018-07-13 #187 @DomBennett
Done!
https://github.com/ropensci/phylotaR
